Use TreeMap for pivot power table
This makes more sense for our use case.

It may be a premature optimization, but I think that we should possibly
use a data structure (custom, most likely, unless we want to use one of
the several-megabyte-large libraries that I know provide it) that allows
for a primative double to be used as a key (not wrapped as an object) so
that the JVM doesn't have to create an object to wrap the primitive
doubles each time we get from the table (reduce work for the garbage
collector, since primitives are kept on the stack). Really just depends on
how often we'll be getting data from the table.